리다이렉트,리스트 뿌리기 완료!

+++ Movie_portfolio/src/main/java/com/portfolio/movie/info/InfoVO.java
@@ -8,11 +8,11 @@ public class InfoVO {
 	private String movieRate;
 	private String openDate;
 	
-	private int star;
-	public int getStar() {
+	private double star;
+	public double getStar() {
 		return star;
 	}
-	public void setStar(int star) {
+	public void setStar(double star) {
 		this.star = star;
 	}
